+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+
+- Print a warning to stderr and fail test if a mock is dropped without having
+  calling `.done()` on it, or if `.done()` is called twice (#59)
+
 ### Fixed
 
 - `Generic` mock: Fix a bug that caused the call to `.done()` to fail if